Python Support and code cleanup

+++ sources/videoio/include/opencv2/videoio.hpp
@@ -69,7 +69,6 @@ typedef struct CvVideoWriter CvVideoWriter;
 
 namespace cv
 {
-#include <vector>
 
 //! @addtogroup videoio
 //! @{
@@ -659,37 +658,37 @@ public:
      */
     CV_WRAP virtual bool open(const String& filename);
 
-	/** @brief Gets Total Number of Devices
+    /** @brief Gets Total Number of Devices
 
     The C function gets the total Number of Devices Connected
      */
-	CV_WRAP virtual bool getDevices(int &devices);
+    CV_WRAP virtual bool getDevices(int &devices);
 
-	/** @brief Gets Specific Device Information
+    /** @brief Gets Specific Device Information
 
     The C function gets the information about the Camera Device such as Device Name, VendorID, ProductID and DevicePath.
      */
-	CV_WRAP virtual bool getDeviceInfo(int index, String &deviceName, String &vid, String &pid, String &devicePath);
+    CV_WRAP virtual bool getDeviceInfo(int index, String &deviceName, String &vid, String &pid, String &devicePath);
 
-	/** @brief Gets Total Number of Formats
+    /** @brief Gets Total Number of Formats
 
     The C function gets the total number of video formats supported by the Camera Device
      */
-	CV_WRAP virtual bool getFormats(int &formats);
+    CV_WRAP virtual bool getFormats(int &formats);
 
-	/** @brief Gets the Video Formats 
+    /** @brief Gets the Video Formats 
 
     The C function also gets the Video resolutions, fps supported by the particular Camera Device
      */
-	CV_WRAP virtual bool getFormatType(int formats, String &formatType, int &width, int &height, int &fps);
+    CV_WRAP virtual bool getFormatType(int formats, String &formatType, int &width, int &height, int &fps);
 
-	/** @brief Sets Video Format to the Camera Device
+    /** @brief Sets Video Format to the Camera Device
 
     The C function also sets the video resolution, fps to the Camera Device
      */
-	CV_WRAP virtual bool setFormatType(int index);
+    CV_WRAP virtual bool setFormatType(int index);
 
-	/** @brief  Open a camera for video capturing
+    /** @brief  Open a camera for video capturing
 
     @overload
 
@@ -802,13 +801,13 @@ public:
      */
     CV_WRAP virtual bool set(int propId, double value);
 
-	/** @brief  Sets a property in the VideoCapture
+    /** @brief  Sets a property in the VideoCapture
 
-	@overload
-
-	Sets the Camera Properties such brightness, contrast, etc., with a particular mode Selection 
-	*/
-	CV_WRAP virtual bool set(int propId, long value, long mode);
+    @overload
+ 
+    Sets the Camera Properties such brightness, contrast, etc., with a particular mode Selection 
+    */
+    CV_WRAP virtual bool set(int propId, double value, double mode);
 
     /** @brief Returns the specified VideoCapture property
 
@@ -829,14 +828,14 @@ public:
     */
     CV_WRAP virtual double get(int propId) const;
 
-	/** @brief  Gets the Camera Properties such as brightness, contrast, hue, etc.,
+    /** @brief  Gets the Camera Properties such as brightness, contrast, hue, etc.,
 	
-	@overload
+    @overload
 
-	Gets the Particular Property's Minimum, Maximum, SupportedMode, CurrentMode, CurrentValue, DefaultValue
-	*/
+    Gets the Particular Property's Minimum, Maximum, SupportedMode, CurrentMode, CurrentValue, DefaultValue
+    */
 
-	CV_WRAP virtual bool get(int propId, long &min, long &max, long &steppingDelta, long &supportedMode, long &currentValue, long &currentMode, long &defaultValue);
+    CV_WRAP virtual bool get(int propId, double &min, double &max, double &steppingDelta, double &supportedMode, double &currentValue, double &currentMode, double &defaultValue);
 
     /** @brief Open video file or a capturing device or a IP video stream for video capturing with API Preference
